+++ app/components/Sidebar.tsx
@@ -57,6 +57,8 @@ function Sidebar() {
     </a>
 </div> */}
 
+
+
 <div className="flex flex-col items-center w-64 h-full overflow-hidden text-gray-400 bg-gray-900 rounded">
     <a className="flex items-center w-full px-3 mt-3" href="#">
         <svg className="w-8 h-8 fill-current" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
